Fix OOM issue for request not freed
in the following commit, it has added code for
	if (prev)
		i915_request_put(prev);

commit 930f3f164964ea823e1fa342d6744d5caec20fd9 (a1)
Author: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Date:   Thu Jul 20 11:35:44 2023 +0200

    drm/i915: Fix premature release of request's reusable memory

but these two lines of code was removed by mistake in commit

commit a30f5dece3420165c5f1c4f225d384c7bca91e23 (b1_5)
Merge: 20090f47f7e0 eb99a642b7e6
Author: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>
Date:   Tue Dec 5 15:57:17 2023 +0530

    Merge commit 'eb99a642b7e6b94282662389ff05737cb8285db9' into dev

    Merged 5.15.138 android13-5.15-lts

It causees the request not be freed, and lead to kernel memory leak.
Fix it by adding back the request free routine.

Test Done:
1. Fast boot flash OK
2. adb reboot OK
3. 24-hours monkey test OK

Tracked-On: OAM-114492
Signed-off-by: Xie, Chao <chao.xie@intel.com>