touchscreen/ef1e_tp: Synchronize with serdes and support power management
There is a race condition in loading of serdes and EF1E touchscreen
driver: if the touchscreen driver starts to read from or write to MCU
before initialization of serdes is complete, terrible things would
happen (e.g., we saw i2c timeout errors suspiciously related with this
sort of race condition).  This is especially likely to happen during
the process of suspending and resuming, where the touchscreen driver
seems to be loaded and started before the serdes driver.

This commit is intended to synchronize between the serdes driver and the
touchscreen driver by deferring the work of the touchscreen driver until
initialization/resetting of serdes gets done.  Implementing power
management callbacks is a good way to achieve this goal.

Test-done:
	Boot Android, suspend and then resume it.  Touch and display is
	working properly.

Tracked-On: OAM-117154
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>