touchscreen/ef1e_tp: Don't install IRQ handler when serdes is not ready
Since we switched to level triggered IRQ, it's observed that when the
serdes screen is not present the overwhelming IRQ would cause kernel
panic.  We suspect this is due to floating state of the GPIO pin.

To cure the problem, we install IRQ handler only when the serdes is
ready and retry for 10 times when it is not.

Test-done:
  Boot Android with no serdes screen.

Tracked-On: OAM-120832
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>