touchscreen/ef1e_tp: Destroy leaked workqueue upon failure
Bump the version to 2024062101.

The workqueue for initialization of interrupt of serdes was not
destroyed upon failure during probing and hence leaked.  In most
situation this is just fine but if we gets really unlucky that the work
bound to the workqueue runs after the probe function returns, kernel
panic is doomed.

Also move creation of the workqueue and queuing of the work to the end
of the probe function to avoid unnecessary work when serdes screen is
not present.

Test-done:
  Boot Android with/without serdes screen.

Tracked-On: OAM-121149
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>